// ui/admin/src/models/pluginCatalogEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wanaku API
 * Wanaku MCP proxy management API
 * OpenAPI spec version: 0.3.0
 */
import type { PluginCatalogEntryMetadata } from "./pluginCatalogEntryMetadata";
import type { PluginRequires } from "./pluginRequires";
import type { ServiceRequirement } from "./serviceRequirement";

/**
 * Metadata for an available plugin and its installation archive.
 */
export interface PluginCatalogEntry {
  dependencies?: ServiceRequirement[];
  description?: string;
  id: string;
  license?: string;
  metadata?: PluginCatalogEntryMetadata;
  name: string;
  publisher?: string;
  requires?: PluginRequires;
  url: string;
  version: string;
}